Rename build-waterline-ontology utility for clarity (besides initializing the Waterline ORM instance, it also runs automigrations).
