+++ lib/build-ontology-and-run-auto-migrations.js
