+++ lib/initialize.js
@@ -10,7 +10,7 @@ var chalk = require('chalk');
 var loadModelsAndCustomAdapters = require('./load-models-and-custom-adapters');
 var validateDatastoreConfig = require('./validate-datastore-config');
 var validateModelDef = require('./validate-model-def');
-var buildWaterlineOntology = require('./build-waterline-ontology');
+var buildOntologyAndRunAutoMigrations = require('./build-ontology-and-run-auto-migrations');
 var loadAdapterFromAppDependencies = require('./load-adapter-from-app-dependencies');
 var buildRegisteredDatastoreInstance = require('./build-registered-datastore-instance');
 
@@ -325,7 +325,7 @@ module.exports = function initialize(hook, sails, done){
         return next(new Error('SAILS EXITING'));
       }
 
-      buildWaterlineOntology(hook, sails, function (err, freshOntology) {
+      buildOntologyAndRunAutoMigrations(hook, sails, function (err, freshOntology) {
         if (err) { return next(err); }
 
         // Finally, continue onward, passing the ontology through for use below.
